LibFix (#13)
* Fixes Scene constructor and the possible memory fault

* Serialization fix

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* NAMSC-41 event tree ( scenery objects not implemented yet) (#12)

* NAMSC-41 - added event tree part 1

* NAMSC-41 - fixed conflicts ( not working )

* NAMSC-41 - connected event tree to properties ( todo: sceneryobjects in event tree )

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Serialization fix

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Serialization and deserialization implemented

* Serialization / deserialization implemented

* Fixes GetType ambiguity, turns Scenery's maps into lists and provides helpers for them. Adds Novel clear method. Changes instanized templates to true templates. Merges with changes already uploaded to the master. Minor bugfixes and QoL

+++ NAMSC_lib/src/Novel/Action/Audio/ActionAudioSetSounds.cpp
@@ -52,9 +52,10 @@ ActionAudioSetSounds::ActionAudioSetSounds(Event* const parentEvent, const Audio
 void ActionAudioSetSounds::serializableLoad(QDataStream& dataStream)
 {
 	ActionAudio::serializableLoad(dataStream);
+
 	uint soundsSize;
 	dataStream >> soundsSize;
-	for (uint i = 0; i != soundsSize; ++i)
+	for (uint i = 0u; i != soundsSize; ++i)
 	{
 		Sound sound;
 		dataStream >> sound;
@@ -66,7 +67,8 @@ void ActionAudioSetSounds::serializableLoad(QDataStream& dataStream)
 void ActionAudioSetSounds::serializableSave(QDataStream& dataStream) const
 {
 	ActionAudio::serializableSave(dataStream);
-	dataStream << sounds.size();
+
+	dataStream << static_cast<uint>(sounds.size());
 	for (const std::pair<const QString, Sound>& sound : sounds)
 		dataStream << sound.second;
 }
